Update old examples
