+++ examples/ResizeExample.js
@@ -6,83 +6,85 @@
 
 import FakeObjectDataListStore from './helpers/FakeObjectDataListStore';
 import { TextCell } from './helpers/cells';
-import { Table, Column, Cell } from 'fixed-data-table-2';
+import { Table, Column, Plugins } from 'fixed-data-table-2';
 import React from 'react';
 
-// Todo: Instead of adding a new example, modify the existing one
+
 class ResizeExample extends React.Component {
-  constructor(props) {
-    super(props);
+    constructor(props) {
+        super(props);
+        this.state = {
+            dataList: new FakeObjectDataListStore(1000000),
+            columnWidths: {
+                firstName: 240,
+                lastName: 150,
+                sentence: 140,
+                companyName: 60,
+            },
+        };
 
-    this.state = {
-      dataList: new FakeObjectDataListStore(1000000),
-      columnWidths: {
-        firstName: 240,
-        lastName: 150,
-        sentence: 140,
-        companyName: 60,
-      },
-    };
+        this._onColumnResizeEndCallback = this._onColumnResizeEndCallback.bind(this);
+    }
 
-    this._onColumnResizeEndCallback = this._onColumnResizeEndCallback.bind(this);
-  }
+    _onColumnResizeEndCallback(newColumnWidth, columnKey) {
+        this.setState(({ columnWidths }) => ({
+            columnWidths: {
+                ...columnWidths,
+                [columnKey]: newColumnWidth,
+            }
+        }));
+    }
 
-  _onColumnResizeEndCallback(newColumnWidth, columnKey) {
-    this.setState(({columnWidths}) => ({
-      columnWidths: {
-        ...columnWidths,
-        [columnKey]: newColumnWidth,
-      }
-    }));
-  }
-
-  render() {
-    var {dataList, columnWidths} = this.state;
-    return (
-      <Table
-        rowHeight={30}
-        headerHeight={50}
-        rowsCount={dataList.getSize()}
-        onColumnResizeEndCallback={this._onColumnResizeEndCallback}
-        isColumnResizing={false}
-        touchScrollEnabled={true}
-        width={1000}
-        height={500}
-        {...this.props}>
-        <Column
-          columnKey="firstName"
-          header={<Cell>First Name</Cell>}
-          cell={<TextCell data={dataList} />}
-          fixed={true}
-          width={columnWidths.firstName}
-          isResizable={true}
-        />
-        <Column
-          columnKey="lastName"
-          header={<Cell>Last Name (min/max constrained)</Cell>}
-          cell={<TextCell data={dataList} />}
-          width={columnWidths.lastName}
-          isResizable={true}
-          minWidth={70}
-          maxWidth={170}
-        />
-        <Column
-          columnKey="companyName"
-          header={<Cell>Company</Cell>}
-          cell={<TextCell data={dataList} />}
-          width={columnWidths.companyName}
-          isResizable={true}
-        />
-        <Column
-          columnKey="sentence"
-          header={<Cell>Sentence</Cell>}
-          cell={<TextCell data={dataList} />}
-          width={columnWidths.sentence}
-          isResizable={true}
-        />
-      </Table>
-    );
-  }
+    render() {
+        var { dataList, columnWidths } = this.state;
+        return (
+            <Table
+                rowHeight={30}
+                headerHeight={50}
+                rowsCount={dataList.getSize()}
+                touchScrollEnabled={true}
+                width={1000}
+                height={500}
+                {...this.props}>
+                <Column
+                    columnKey="firstName"
+                    header={<Plugins.ResizeReorderCell
+                        onColumnResizeEndCallback={this._onColumnResizeEndCallback}>
+                        First Name </Plugins.ResizeReorderCell>}
+                    cell={<TextCell data={dataList}/>}
+                    fixed={true}
+                    width={columnWidths.firstName}
+                />
+                <Column
+                    columnKey="lastName"
+                    header={<Plugins.ResizeReorderCell
+                        minWidth={70}
+                        maxWidth={170}
+                        onColumnResizeEndCallback={this._onColumnResizeEndCallback}>
+                        Last Name (min/max constrained)
+                    </Plugins.ResizeReorderCell>}
+                    cell={<TextCell data={dataList}/>}
+                    width={columnWidths.lastName}
+                />
+                <Column
+                    columnKey="companyName"
+                    header={<Plugins.ResizeReorderCell
+                        onColumnResizeEndCallback={this._onColumnResizeEndCallback}>
+                        Company </Plugins.ResizeReorderCell>}
+                    cell={<TextCell data={dataList}/>}
+                    width={columnWidths.companyName}
+                />
+                <Column
+                    columnKey="sentence"
+                    header={<Plugins.ResizeReorderCell
+                        onColumnResizeEndCallback={this._onColumnResizeEndCallback}>
+                        Sentence </Plugins.ResizeReorderCell>}
+                    cell={<TextCell data={dataList}/>}
+                    width={columnWidths.sentence}
+                />
+            </Table>
+        );
+    }
 }
 
 export default ResizeExample;
